# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admission Route API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/me/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendations",
        "operationId": "get_recommendations_me_recommendations_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/preview": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Preview",
        "description": "Stateless what-if: never writes anything.",
        "operationId": "preview_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/changes/latest": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Latest Changes",
        "operationId": "latest_changes_me_changes_latest_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestChanges"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/roadmap": {
      "get": {
        "tags": [
          "roadmap"
        ],
        "summary": "Get Roadmap",
        "operationId": "get_roadmap_me_roadmap_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/roadmap/steps/{step_id}": {
      "patch": {
        "tags": [
          "roadmap"
        ],
        "summary": "Patch Step",
        "operationId": "patch_step_me_roadmap_steps__step_id__patch",
        "parameters": [
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Step Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roadmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/roadmap.ics": {
      "get": {
        "tags": [
          "roadmap"
        ],
        "summary": "Roadmap Ics",
        "operationId": "roadmap_ics_me_roadmap_ics_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/favorites": {
      "get": {
        "tags": [
          "roadmap"
        ],
        "summary": "Get Favorites",
        "operationId": "get_favorites_me_favorites_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Favorites Me Favorites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/favorites/{university_id}": {
      "put": {
        "tags": [
          "roadmap"
        ],
        "summary": "Add Favorite",
        "operationId": "add_favorite_me_favorites__university_id__put",
        "parameters": [
          {
            "name": "university_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "University Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "roadmap"
        ],
        "summary": "Remove Favorite",
        "operationId": "remove_favorite_me_favorites__university_id__delete",
        "parameters": [
          {
            "name": "university_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "University Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Achievement": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "SAT",
              "IELTS",
              "TOEFL",
              "OLYMPIAD",
              "PROJECT",
              "VOLUNTEER",
              "COMPETITION",
              "OTHER"
            ],
            "title": "Type"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "school",
                  "city",
                  "national",
                  "international"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "done",
              "planned"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "date",
          "status"
        ],
        "title": "Achievement"
      },
...
      "ChanceChange": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "from": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "From"
          },
          "to": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "id",
          "from",
          "to"
        ],
        "title": "ChanceChange"
      },
...
      "ComputeResponse": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/RecommendationResult"
          },
          "roadmap": {
            "$ref": "#/components/schemas/Roadmap"
          },
          "diff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Diff"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "result",
          "roadmap",
          "diff"
        ],
        "title": "ComputeResponse"
      },
      "Conflict": {
        "properties": {
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "step_id",
          "message"
        ],
        "title": "Conflict"
      },
...
      "Diff": {
        "properties": {
          "added": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Added"
          },
          "removed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed"
          },
          "tier_changed": {
            "items": {
              "$ref": "#/components/schemas/TierChange"
            },
            "type": "array",
            "title": "Tier Changed"
          },
          "chance_changed": {
            "items": {
              "$ref": "#/components/schemas/ChanceChange"
            },
            "type": "array",
            "title": "Chance Changed"
          },
          "gaps_closed": {
            "items": {
              "$ref": "#/components/schemas/GapClosed"
            },
            "type": "array",
            "title": "Gaps Closed"
          },
          "roadmap_added": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roadmap Added"
          },
          "roadmap_removed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roadmap Removed"
          },
          "cause": {
            "type": "string",
            "title": "Cause"
          }
        },
        "type": "object",
        "required": [
          "added",
          "removed",
          "tier_changed",
          "chance_changed",
          "gaps_closed",
          "roadmap_added",
          "roadmap_removed",
          "cause"
        ],
        "title": "Diff"
      },
      "Excluded": {
        "properties": {
          "university_id": {
            "type": "string",
            "title": "University Id"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "university_id",
          "reasons"
        ],
        "title": "Excluded"
      },
...
      "GapClosed": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code"
        ],
        "title": "GapClosed"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LatestChanges": {
        "properties": {
          "diff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Diff"
              },
              {
                "type": "null"
              }
            ]
          },
          "cause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cause"
          },
          "at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "diff",
          "cause",
          "at"
        ],
        "title": "LatestChanges"
      },
...
      "PreviewIn": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "priorities_override": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Priorities"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "PreviewIn"
      },
      "Priorities": {
        "properties": {
          "cost": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Cost",
            "default": 0.5
          },
          "prestige": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Prestige",
            "default": 0.5
          },
          "location": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Location",
            "default": 0.5
          },
          "aid": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Aid",
            "default": 0.5
          }
        },
        "type": "object",
        "title": "Priorities"
      },
      "Profile": {
        "properties": {
          "grade": {
            "type": "integer",
            "enum": [
              10,
              11,
              12
            ],
            "title": "Grade"
          },
          "gpa5": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 2.0,
            "title": "Gpa5"
          },
          "majors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Majors"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Countries"
          },
          "budget_per_year_usd": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Budget Per Year Usd"
          },
          "needs_aid": {
            "type": "boolean",
            "title": "Needs Aid"
          },
          "intake_year": {
            "type": "integer",
            "title": "Intake Year"
          },
          "priorities": {
            "$ref": "#/components/schemas/Priorities"
          },
          "achievements": {
            "items": {
              "$ref": "#/components/schemas/Achievement"
            },
            "type": "array",
            "title": "Achievements",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "grade",
          "gpa5",
          "majors",
          "countries",
          "budget_per_year_usd",
          "needs_aid",
          "intake_year",
          "priorities",
          "created_at"
        ],
        "title": "Profile"
      },
...
      "Reason": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "plus",
              "risk",
              "blocker"
            ],
            "title": "Kind"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "profile_field": {
            "type": "string",
            "title": "Profile Field"
          },
          "gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "code",
          "text",
          "profile_field"
        ],
        "title": "Reason"
      },
      "Recommendation": {
        "properties": {
          "university_id": {
            "type": "string",
            "title": "University Id"
          },
          "tier": {
            "type": "string",
            "enum": [
              "dream",
              "target",
              "safety"
            ],
            "title": "Tier"
          },
          "chance": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Chance"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "university_id",
          "tier",
          "chance",
          "score",
          "reasons"
        ],
        "title": "Recommendation"
      },
      "RecommendationResult": {
        "properties": {
          "recs": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recs"
          },
          "excluded": {
            "items": {
              "$ref": "#/components/schemas/Excluded"
            },
            "type": "array",
            "title": "Excluded"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "computed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Computed At"
          }
        },
        "type": "object",
        "required": [
          "recs",
          "excluded",
          "suggestions",
          "computed_at"
        ],
        "title": "RecommendationResult"
      },
      "Roadmap": {
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RoadmapStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/Conflict"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "next_step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Step Id"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "steps",
          "conflicts",
          "next_step_id",
          "progress"
        ],
        "title": "Roadmap"
      },
      "RoadmapStep": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "exam",
              "document",
              "academic",
              "activity",
              "application"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "university_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "University Ids"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "done": {
            "type": "boolean",
            "title": "Done"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "due_date",
          "depends_on",
          "university_ids",
          "is_demo",
          "done",
          "priority"
        ],
        "title": "RoadmapStep"
      },
...
      "StepPatch": {
        "properties": {
          "done": {
            "type": "boolean",
            "title": "Done"
          }
        },
        "type": "object",
        "required": [
          "done"
        ],
        "title": "StepPatch"
      },
...
      "TierChange": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "from": {
            "type": "string",
            "enum": [
              "dream",
              "target",
              "safety"
            ],
            "title": "From"
          },
          "to": {
            "type": "string",
            "enum": [
              "dream",
              "target",
              "safety"
            ],
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "id",
          "from",
          "to"
        ],
        "title": "TierChange"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}